+++ changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [5.1.1 - 2020-04-15]
 ### Changed
 - no cache control for requests
+- avoid using header parameters
 
 ## [5.1.0 - 2020-03-31]
 ### Added